+++ objects/crafting/liquidpumpwell/liquidpumpwell.object
@@ -24,7 +24,7 @@
     }
   ],
     "outputNodes" : [ [2, 1] ],
-  "scripts" : [ "/scripts/well.lua" ],
+  "scripts" : [ "/scripts/fuwell.lua" ],
   "scriptDelta" : 10,
   "slotCount" : 3,
   "uiConfig" : "/interface/chests/chest<slots>.config",
