+++ objects/crafting/xiwell/xiwell.object
@@ -24,7 +24,7 @@
     }
   ],
     "outputNodes" : [ [2, 1] ],
-  "scripts" : [ "/scripts/well.lua" ],
+  "scripts" : [ "/scripts/fuwell.lua" ],
   "scriptDelta" : 260,
   "slotCount" : 1,
   "uiConfig" : "/interface/chests/chest<slots>.config",
