+++ objects/novakid/frontierwatertower/frontierwatertower.object
@@ -23,7 +23,7 @@
     }
   ],
   "outputNodes" : [ [2, 1] ],
-  "scripts" : [ "/scripts/well.lua" ],
+  "scripts" : [ "/scripts/fuwell.lua" ],
   "scriptDelta" : 320,
   "slotCount" : 1,
   "uiConfig" : "/interface/chests/chest<slots>.config",
